ditto

+++ src/net/ftb/data/UserManager.java
@@ -47,18 +47,21 @@ public class UserManager {
     }
 
     public void write () throws IOException {
-        try {
-            if(_oldFile.exists()) {
-                _oldFile.delete();
+
+        if (OSUtils.getCurrentOS() == OSUtils.OS.WINDOWS) {
+            try {
+                if (_oldFile.exists()) {
+                    _oldFile.delete();
+                }
+
+                if (_file.exists()) {
+                    _file.delete();
+                }
+            } finally {
+
             }
-            
-            if(_file.exists()) {
-                _file.delete();
-            }
-        } finally {
-            
         }
-        
+    
         FileOutputStream fileOutputStream = new FileOutputStream(_file);
         ObjectOutputStream objectOutputStream = new ObjectOutputStream(fileOutputStream);
         try {
